+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed:
  - Pins *activity()* and *tried()* methods always return Pagination object.
+ - Boards *forUser* method returns empty array for no boards.
 
 ## [5.2.3] - 2017-01-24
 ### Fixed: